t/collectio-functions/funciones-sobre-colecciones closes #20
