--- examples/collection-functions/collection-functions.sh
@@ -1,7 +1,0 @@
-$ go run collection-functions.go 
-2
-false
-true
-false
-[peach apple pear]
-[PEACH APPLE PEAR PLUM]
