+++ examples/funciones-sobre-colecciones/funciones-sobre-colecciones.sh
@@ -1,0 +1,7 @@
+$go run funciones-sobre-colecciones.go
+2
+false
+true
+false
+[pera ciruela]
+[DURAZNO MANZANA PERA CIRUELA]
