+++ public/defer
@@ -181,7 +181,7 @@ se cierra después de escribir en él</p>
       
       
       <p class="next">
-        Siguiente ejemplo: <a href="collection-functions">Collection Functions</a>.
+        Siguiente ejemplo: <a href="funciones-sobre-colecciones">Funciones sobre colecciones</a>.
       </p>
       
       <p class="footer">
