+++ public/index.html
@@ -119,7 +119,7 @@
       
         <li><a href="defer">Defer</a></li>
       
-        <li><a href="collection-functions">Collection Functions</a></li>
+        <li><a href="funciones-sobre-colecciones">Funciones sobre colecciones</a></li>
       
         <li><a href="funciones-de-cadenas">Funciones de Cadenas</a></li>
